Revert "Revert tests module refactor"
This reverts commit bc133c99bd3a0caf29ac76fd5cc4d3eef764ba2b.
